🔧fix: Improved error messageds

+++ src-tauri/src/file_manager.rs
@@ -1,5 +1,5 @@
-use std::{fs, path::Path};
 use std::path::PathBuf;
+use std::{fs, path::Path};
 use tauri::{AppHandle, Manager};
 
 use crate::cbz_viewer::{CbzViewer, ComicInfo};
@@ -11,8 +11,12 @@ pub struct FileManager {
 
 impl FileManager {
     pub fn new(app: &AppHandle) -> Result<Self, String> {
-        let mut data_dir = app.path().app_data_dir().map_err(|e| e.to_string()).unwrap();
-        
+        let mut data_dir = app
+            .path()
+            .app_data_dir()
+            .map_err(|e| e.to_string())
+            .unwrap();
+
         fs::create_dir_all(&data_dir).map_err(|e| e.to_string())?;
 
         data_dir.push("comics");
@@ -25,48 +29,69 @@ impl FileManager {
 
     pub fn get_full_path(&self, file_name: &str) -> Result<PathBuf, String> {
         let file = Path::new(file_name);
-        let file_stem = file.file_stem().ok_or("Invalid file name")?.to_str().ok_or("Invalid file name")?;
+        let file_stem = file
+            .file_stem()
+            .ok_or("Invalid file name")?
+            .to_str()
+            .ok_or("Invalid file name")?;
 
         Ok(self.directory.join(file_stem).join(file_name))
     }
 
     pub fn add_file(&self, source_path: &str) -> Result<(), String> {
         let source = Path::new(source_path);
-        
-        let file_stem = source.file_stem().ok_or("Invalid file name")?.to_str().ok_or("Invalid file name")?;
-        let file_name = source.file_name().ok_or("Invalid file name")?.to_str().ok_or("Invalid file name")?;
+
+        let file_stem = source
+            .file_stem()
+            .ok_or("Invalid file name")?
+            .to_str()
+            .ok_or("Invalid file name")?;
+        let file_name = source
+            .file_name()
+            .ok_or("Invalid file name")?
+            .to_str()
+            .ok_or("Invalid file name")?;
 
         let folder_path = self.directory.join(file_stem);
         fs::create_dir_all(&folder_path).map_err(|e| e.to_string())?;
 
         let destination_path = folder_path.join(file_name);
+
+        if destination_path.exists() {
+            return Err("File already exists in the library".to_string());
+        }
+
         fs::copy(source_path, &destination_path).map_err(|e| e.to_string())?;
-        
-        let comic_info = match CbzViewer::read_comic_info(destination_path.to_str().ok_or("Invalid path")?) {
-            Ok(info) => info,
-            Err(_) => {
-                println!("Warning: Could not read ComicInfo.xml, using placeholder data.");
-                ComicInfo {
-                    title: file_stem.to_string(),
-                    series: "".to_string(),
-                    number: "".to_string(),
-                    volume: "".to_string(),
-                    summary: "".to_string(),
-                    year: "".to_string(),
-                    month: "".to_string(),
-                    day: "".to_string(),
-                    writer: "Unknown".to_string(),
-                    publisher: "".to_string(),
-                    page_count: "".to_string(),
+
+        let comic_info =
+            match CbzViewer::read_comic_info(destination_path.to_str().ok_or("Invalid path")?) {
+                Ok(info) => info,
+                Err(_) => {
+                    println!("Warning: Could not read ComicInfo.xml, using placeholder data.");
+                    ComicInfo {
+                        title: file_stem.to_string(),
+                        series: "".to_string(),
+                        number: "".to_string(),
+                        volume: "".to_string(),
+                        summary: "".to_string(),
+                        year: "".to_string(),
+                        month: "".to_string(),
+                        day: "".to_string(),
+                        writer: "Unknown".to_string(),
+                        publisher: "".to_string(),
+                        page_count: "".to_string(),
+                    }
                 }
-            },
-        };
+            };
 
-        self.create_metadata_file(&folder_path, &comic_info).map_err(|e| e.to_string())?;
+        self.create_metadata_file(&folder_path, &comic_info)
+            .map_err(|e| e.to_string())?;
 
-        let comic_cover = CbzViewer::extract_cover_image(destination_path.to_str().ok_or("Invalid path")?)?;
+        let comic_cover =
+            CbzViewer::extract_cover_image(destination_path.to_str().ok_or("Invalid path")?)?;
         if let Some(cover_image_data) = comic_cover {
-            self.copy_cover_image(&folder_path, &cover_image_data).map_err(|e| e.to_string())?;
+            self.copy_cover_image(&folder_path, &cover_image_data)
+                .map_err(|e| e.to_string())?;
         }
 
         Ok(())
@@ -74,7 +99,11 @@ impl FileManager {
 
     pub fn delete_file(&self, file_name: &str) -> Result<(), String> {
         let file = Path::new(file_name);
-        let file_stem = file.file_stem().ok_or("Invalid file name")?.to_str().ok_or("Invalid file name")?;
+        let file_stem = file
+            .file_stem()
+            .ok_or("Invalid file name")?
+            .to_str()
+            .ok_or("Invalid file name")?;
 
         let folder_path = self.directory.join(file_stem);
         if folder_path.exists() {
@@ -84,7 +113,11 @@ impl FileManager {
         Ok(())
     }
 
-    pub fn create_metadata_file(&self, folder_path: &PathBuf, comic_info: &ComicInfo) -> Result<(), String> {
+    pub fn create_metadata_file(
+        &self,
+        folder_path: &PathBuf,
+        comic_info: &ComicInfo,
+    ) -> Result<(), String> {
         let metadata_path = folder_path.join("metadata.json");
         let metadata = serde_json::to_string_pretty(comic_info).map_err(|e| e.to_string())?;
         fs::write(metadata_path, metadata).map_err(|e| e.to_string())?;
@@ -92,9 +125,17 @@ impl FileManager {
         Ok(())
     }
 
-    pub fn edit_metadata_file(&self, cbz_path: &String, comic_info: &ComicInfo) -> Result<(), String> {
+    pub fn edit_metadata_file(
+        &self,
+        cbz_path: &String,
+        comic_info: &ComicInfo,
+    ) -> Result<(), String> {
         let file = Path::new(cbz_path);
-        let file_stem = file.file_stem().ok_or("Invalid file name")?.to_str().ok_or("Invalid file name")?;
+        let file_stem = file
+            .file_stem()
+            .ok_or("Invalid file name")?
+            .to_str()
+            .ok_or("Invalid file name")?;
         let metadata_path = self.directory.join(file_stem).join("metadata.json");
         if metadata_path.exists() {
             let metadata = serde_json::to_string_pretty(comic_info).map_err(|e| e.to_string())?;
@@ -105,10 +146,19 @@ impl FileManager {
         }
     }
 
-    pub fn copy_cover_image(&self, folder_path: &PathBuf, cover_image_data: &str) -> Result<(), String> {
+    pub fn copy_cover_image(
+        &self,
+        folder_path: &PathBuf,
+        cover_image_data: &str,
+    ) -> Result<(), String> {
         let cover_path = folder_path.join("cover");
-        let base64_data = cover_image_data.split(',').nth(1).ok_or("Invalid image data")?;
-        let image_data = general_purpose::STANDARD.decode(base64_data).map_err(|e| e.to_string())?;
+        let base64_data = cover_image_data
+            .split(',')
+            .nth(1)
+            .ok_or("Invalid image data")?;
+        let image_data = general_purpose::STANDARD
+            .decode(base64_data)
+            .map_err(|e| e.to_string())?;
 
         let extension = if cover_image_data.starts_with("data:image/jpeg") {
             "jpg"
@@ -117,7 +167,7 @@ impl FileManager {
         } else {
             return Err("Unsupported image format".to_string());
         };
-        
+
         let cover_path = cover_path.with_extension(extension);
         fs::write(cover_path, image_data).map_err(|e| e.to_string())?;
         Ok(())
@@ -141,7 +191,12 @@ impl FileManager {
                 for sub_entry in sub_entries {
                     let sub_entry = sub_entry.map_err(|e| e.to_string())?;
                     let sub_path = sub_entry.path();
-                    if sub_path.is_file() && sub_path.extension().and_then(|s| s.to_str()) == Some("cbz") {
+                    if sub_path.is_file()
+                        && sub_path.extension().and_then(|s| s.to_str()) == Some("cbz")
+                        || sub_path.extension().and_then(|s| s.to_str()) == Some("zip")
+                        || sub_path.extension().and_then(|s| s.to_str()) == Some("cbr")
+                        || sub_path.extension().and_then(|s| s.to_str()) == Some("rar")
+                    {
                         if let Some(file_name) = sub_path.file_name().and_then(|s| s.to_str()) {
                             files.push(file_name.to_string());
                         }
